fix(dynamic-frontend): switch tech-radar plugin to dynamic
