+++ .changeset/silver-jokes-smell.md
@@ -1,0 +1,6 @@
+---
+'backstage-plugin-tech-radar': minor
+'app': minor
+---
+
+Switch Tech Radar plugin to dynamic frontend plugin implementation
