+++ app-config.dynamic-plugins.yaml
@@ -377,3 +377,19 @@ dynamicPlugins:
             props:
               name: Documentation
               icon: docs
+    backstage.plugin-tech-radar:
+      apiFactories:
+        - importName: TechRadarApi
+      appIcons:
+        - name: techRadar
+          importName: TechRadarIcon
+      dynamicRoutes:
+        - path: /tech-radar
+          importName: TechRadarPage
+          menuItem:
+            icon: techRadar
+            text: Tech Radar
+          config:
+            props:
+              width: 1500
+              height: 800
