+++ dynamic-plugins/wrappers/backstage-plugin-tech-radar/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "backstage-plugin-tech-radar",
+  "version": "0.6.9",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "frontend-plugin"
+  },
+  "scripts": {
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test --passWithNoTests --coverage",
+    "clean": "backstage-cli package clean",
+    "export-dynamic": "janus-cli package export-dynamic-plugin"
+  },
+  "dependencies": {
+    "@backstage/core-plugin-api": "1.7.0",
+    "@backstage/plugin-tech-radar": "0.6.9",
+    "@mui/icons-material": "5.14.16"
+  },
+  "devDependencies": {
+    "@backstage/cli": "0.23.1",
+    "@janus-idp/cli": "1.4.3"
+  },
+  "files": [
+    "dist",
+    "dist-scalprum"
+  ],
+  "scalprum": {
+    "name": "backstage.plugin-tech-radar",
+    "exposedModules": {
+      "PluginRoot": "./src/index.ts"
+    }
+  }
+}
